Rename Genymotion-specific things to generalized (Genymotion)Player -> DeviceRenderer GenymotionManager -> DeviceRendererFactory (Genymotion)Instance -> (Device)Renderer
