+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "@genymotion/device-web-player",
   "version": "2.1.3",
-  "description": "Genymotion Virtual Device Web Player",
-  "main": "dist/js/gm-player.min.js",
+  "description": "Genymotion Virtual Device Web Renderer",
+  "main": "dist/js/device-renderer.min.js",
   "files": [
     "dist"
   ],
