+++ src/DeviceRendererFactory.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const GenymotionInstance = require('./GenymotionInstance');
+const DeviceRenderer = require('./DeviceRenderer');
 const defaultsDeep = require('lodash/defaultsDeep');
 
 // Plugins
@@ -28,12 +28,12 @@ const log = require('loglevel');
 log.setDefaultLevel('debug');
 
 // Templates are loaded dynamically from the `templates` folder
-const TEMPLATE_JS = 'genymotion-js';
-const TEMPLATE_CSS = 'genymotion-css';
+const TEMPLATE_JS = 'device-renderer-js';
+const TEMPLATE_CSS = 'device-renderer-css';
 
 // Default options
 const defaultOptions = {
-    template: 'player',
+    template: 'renderer',
     touch: true,
     mouse: true,
     volume: true,
@@ -79,9 +79,9 @@ const defaultOptions = {
 };
 
 /**
- * Setup & create instances of the Genymotion player
+ * Setup & create instances of the device renderer
  */
-module.exports = class GenymotionManager {
+module.exports = class DeviceRendererFactory {
     constructor() {
         this.instances = [];
         /* global GEN_TEMPLATES */
@@ -89,12 +89,12 @@ module.exports = class GenymotionManager {
     }
 
     /**
-     * Setup a device player instance in the given dom element, for the Genymotion Cloud instance identified by its instanceWebRTCUrl.
+     * Setup a device renderer instance in the given dom element, for the device instance identified by its instanceWebRTCUrl.
      *
-     * @param  {HTMLElement|string} dom                            The DOM element (or its ID) to setup the device player into.
+     * @param  {HTMLElement|string} dom                            The DOM element (or its ID) to setup the device renderer into.
      * @param  {string}             webRTCUrl                      WebRTC URL of the instance.
      * @param  {Object}             options                        Various configuration options.
-     * @param  {string}             options.template               Template to use. Default: 'player'.
+     * @param  {string}             options.template               Template to use. Default: 'renderer'.
      * @param  {boolean}            options.touch                  Touch support activated. Default: true.
      * @param  {boolean}            options.mouse                  Mouse support activated. Default: true.
      * @param  {boolean}            options.volume                 Audio volume control support activated. Default: true.
@@ -131,9 +131,9 @@ module.exports = class GenymotionManager {
      * @param  {string}             options.turn.username          WebRTC TURN servers username.
      * @param  {string}             options.turn.credential        WebRTC TURN servers password.
      * @param  {boolean}            options.turn.default           Whether or not we should use the TURN servers by default. Default: false.
-     * @return {GenymotionInstance}                                The Genymotion instance.
+     * @return {DeviceRenderer}                                    The device renderer instance.
      */
-    setupInstance(dom, webRTCUrl, options) {
+    setupRenderer(dom, webRTCUrl, options) {
         if (typeof dom === 'string') {
             dom = document.getElementById(dom);
         }
@@ -147,7 +147,7 @@ module.exports = class GenymotionManager {
         options.buttons = options.volume || options.rotation || options.navbar || options.power;
 
         log.debug('Creating genymotion display on ' + webRTCUrl);
-        dom.classList.add('gm-player-instance');
+        dom.classList.add('device-renderer-instance');
         dom.classList.add('gm-template-' + options.template);
         document.body.classList.add('gm-template-' + options.template + '-body');
 
@@ -157,10 +157,10 @@ module.exports = class GenymotionManager {
     /**
      * Loads the selected template.
      *
-     * @param  {HTMLElement}        dom       The DOM element to setup the device player into.
+     * @param  {HTMLElement}        dom       The DOM element to setup the device renderer into.
      * @param  {string}             template  Template to use.
      * @param  {Object}             options   Various configuration options.
-     * @return {GenymotionInstance}          The Genymotion instance.
+     * @return {DeviceRenderer}               The device renderer instance.
      */
     loadTemplate(dom, template, options) {
         const head = document.getElementsByTagName('head')[0];
@@ -193,12 +193,12 @@ module.exports = class GenymotionManager {
     /**
      * Initialize all the needed plugins.
      *
-     * @param  {HTMLElement}        dom      The DOM element to setup the device player into.
+     * @param  {HTMLElement}        dom      The DOM element to setup the device renderer into.
      * @param  {Object}             options  Various configuration options.
-     * @return {GenymotionInstance}          The Genymotion instance.
+     * @return {DeviceRenderer}              The device renderer instance.
      */
     addPlugins(dom, options) {
-        const instance = new GenymotionInstance(dom, options);
+        const instance = new DeviceRenderer(dom, options);
         this.instances.push(instance);
 
         const pluginInitMap = [
