+++ src/plugins/Fullscreen.js
@@ -2,7 +2,7 @@
 
 /**
  * Display fullscreen plugin.
- * Gives the ability to display the player on the whole screen.
+ * Gives the ability to display the renderer on the whole screen.
  */
 module.exports = class Fullscreen {
     /**
@@ -25,7 +25,7 @@ module.exports = class Fullscreen {
     }
 
     /**
-     * Add the button to the player toolbar.
+     * Add the button to the renderer toolbar.
      */
     renderToolbarButton() {
         const toolbars = this.instance.getChildByClass(this.instance.root, 'gm-toolbar');
