+++ src/plugins/IOThrottling.js
@@ -46,7 +46,7 @@ module.exports = class IOThrottling extends OverlayPlugin {
     }
 
     /**
-     * Add the button to the player toolbar.
+     * Add the button to the renderer toolbar.
      */
     renderToolbarButton() {
         const toolbars = this.instance.getChildByClass(this.instance.root, 'gm-toolbar');
