+++ src/plugins/StreamBitrate.js
@@ -26,7 +26,7 @@ module.exports = class StreamBitrate {
     }
 
     /**
-     * Add the button to the player toolbar.
+     * Add the button to the renderer toolbar.
      */
     renderToolbarButton() {
         const toolbars = this.instance.getChildByClass(this.instance.root, 'gm-toolbar');
