+++ src/scss/base/_genymotion.scss
@@ -1,7 +1,7 @@
 /**
- * Common player style
+ * Common renderer style
  */
-.gm-player-instance {
+.device-renderer-instance {
   text-align: center;
 
   *, *::after, *::before {
