+++ src/scss/components/_battery.scss
@@ -1,7 +1,7 @@
 /**
  * Battery plugin styles
  */
-.gm-player-instance .gm-battery-plugin {
+.device-renderer-instance .gm-battery-plugin {
   .gm {
     &-charging-group {
       margin-top: 15px;
