+++ src/scss/components/_clipboard.scss
@@ -1,7 +1,7 @@
 /**
  * Clipboard plugin styles
  */
-.gm-player-instance .gm-clipboard-plugin {
+.device-renderer-instance .gm-clipboard-plugin {
   .gm {
     &-clipboard-input {
       min-height: 300px;
