+++ src/scss/components/_fullscreen.scss
@@ -1,7 +1,7 @@
 /**
  * Fullscreen plugin styles
  */
-.gm-player-instance .gm-fullscreen {
+.device-renderer-instance .gm-fullscreen {
   position: absolute;
   top: 0;
   left: 0;
@@ -10,7 +10,7 @@
   padding: 0;
 
   .gm {
-    &-player {
+    &-renderer {
       background: $black !important;
       height: 100%;
       max-height: 100% !important;
