+++ src/scss/components/_gps.scss
@@ -1,7 +1,7 @@
 /**
  * GPS plugin styles
  */
-.gm-player-instance .gm-gps {
+.device-renderer-instance .gm-gps {
   &-controls {
     label {
       margin-top: 10px;
