+++ src/scss/components/_imei.scss
@@ -1,7 +1,7 @@
 /**
  * Identifiers (IMEI) plugin styles
  */
-.gm-player-instance .gm-identifiers-plugin {
+.device-renderer-instance .gm-identifiers-plugin {
   .gm-identifier-device, .gm-identifier-android {
     .gm-description {
       font-size: 90%;
