+++ src/scss/components/_iothrottling.scss
@@ -1,7 +1,7 @@
 /**
  * IO Throttling plugin styles
  */
-.gm-player-instance .gm-iothrottling-plugin {
+.device-renderer-instance .gm-iothrottling-plugin {
   .gm-inputs {
     margin-bottom: 15px;
