+++ src/scss/components/_network.scss
@@ -1,7 +1,7 @@
 /**
  * Network plugin styles
  */
-.gm-player-instance .gm-network-plugin {
+.device-renderer-instance .gm-network-plugin {
   .gm {
     &-checkbox-group {
       margin-bottom: 15px;
