+++ src/scss/components/_phone.scss
@@ -1,7 +1,7 @@
 /**
  * Phone plugin styles
  */
-.gm-player-instance .gm-phone-plugin {
+.device-renderer-instance .gm-phone-plugin {
   .gm-phone-group {
     &:first-child {
       margin-bottom: 12px;
