+++ src/scss/components/_screencast.scss
@@ -1,7 +1,7 @@
 /**
  * Screencast plugin styles
  */
-.gm-player-instance {
+.device-renderer-instance {
   .gm-screencast-plugin {
     width: 400px;
     left: calc(50% - 200px);
