+++ src/scss/components/_toolbar.scss
@@ -1,7 +1,7 @@
 /**
- * Player toolbar styles
+ * renderer toolbar styles
  */
-.gm-player-instance  {
+.device-renderer-instance  {
   .gm-toolbar img:hover,
   .gm-streamrate-chooser img:hover {
       cursor: pointer;
