+++ src/scss/components/_turn.scss
@@ -2,7 +2,7 @@
  * Turn server plugin styles
  */
 
-.gm-player-instance .gm-default-turn-button {
+.device-renderer-instance .gm-default-turn-button {
   text-decoration: none;
 
   .gm-default-turn-used {
