+++ src/scss/components/_upload.scss
@@ -1,7 +1,7 @@
 /**
  * Upload plugin styles
  */
-.gm-player-instance {
+.device-renderer-instance {
   .gm {
     &-wrapper {
       display: inline-block;
