+++ src/scss/components/_uploader.scss
@@ -1,7 +1,7 @@
 /**
  * File upload plugin styles
  */
-.gm-player-instance .gm-uploader-content {
+.device-renderer-instance .gm-uploader-content {
   .gm {
     &-upload-icon {
       display: inline-block;
