+++ src/templates/renderer/dom.html
@@ -1,5 +1,5 @@
 <div class="gm-wrapper">
-    <div class="gm-player">
+    <div class="device-renderer">
         <div class="gm-video-wrapper">
             <video class="gm-video" autoplay preload="none">
                 Your browser does not support the VIDEO tag
