+++ src/templates/renderer_no_toolbar/dom.html
@@ -1,6 +1,6 @@
 <div class="gm-wrapper">
     <div class="gm-video-wrapper">
-        <video class="gm-video" autoplay poster="img/screenshot.png">
+        <video class="gm-video" autoplay poster="src/assets/images/loader.svg">
             Your browser does not support the VIDEO tag
         </video>
     </div>
